LayoutTests/css3/scroll-snap/scroll-snap-property-parsing-expected.txt

Test the parsing of the -webkit-scroll-snap-* properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



mandatory type : mandatory
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-type') is 'mandatory'

proximity type : proximity
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-type') is 'proximity'

element points along x axis : elements
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-points-x') is 'elements'

percentage points along x axis : 100% 50%
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-points-x') is '100% 50%'

pixel points along x axis : 100px 50px
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-points-x') is '100px 50px'

percentage points repeat along x axis : repeat(100%)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-points-x') is 'repeat(100%)'

pixe points repeat along x axis : repeat(25px)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-points-x') is 'repeat(25px)'

percentage points along x axis with percentage repeat : 100% repeat(100%)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-points-x') is '100% repeat(100%)'

pixel points along x axis with percentage repeat : 100px 50px repeat(25%)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-points-x') is '100px 50px repeat(25%)'

percentage points along x axis with pixel repeat : 100% 50% repeat(40px)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-points-x') is '100% 50% repeat(40px)'

pixel points along x axis with pixel repeat : 100px repeat(42px)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-points-x') is '100px repeat(42px)'

element points along y axis : elements
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-points-y') is 'elements'

percentage points along y axis : 100% 50%
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-points-y') is '100% 50%'

pixel points along y axis : 100px 50px
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-points-y') is '100px 50px'

percentage points repeat along y axis : repeat(100%)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-points-y') is 'repeat(100%)'

pixe points repeat along y axis : repeat(25px)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-points-y') is 'repeat(25px)'

percentage points along y axis with percentage repeat : 100% repeat(100%)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-points-y') is '100% repeat(100%)'

pixel points along y axis with percentage repeat : 100px 50px repeat(25%)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-points-y') is '100px 50px repeat(25%)'

percentage points along y axis with pixel repeat : 100% 50% repeat(40px)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-points-y') is '100% 50% repeat(40px)'

pixel points along y axis with pixel repeat : 100px repeat(42px)
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-points-y') is '100px repeat(42px)'

em points along x axis with pixel repeat : 100em repeat(42em)
PASS cssRule.type is 1
PASS declaration.length is 1
FAIL declaration.getPropertyValue('-webkit-scroll-snap-points-x') should be 100em repeat 42em. Was 100em repeat(42em).

mm along x axis with pixel repeat : 100mm repeat(42mm)
PASS cssRule.type is 1
PASS declaration.length is 1
FAIL declaration.getPropertyValue('-webkit-scroll-snap-points-x') should be 100mm repeat 42mm. Was 100mm repeat(42mm).

in along x axis with pixel repeat : 100in repeat(42in)
PASS cssRule.type is 1
PASS declaration.length is 1
FAIL declaration.getPropertyValue('-webkit-scroll-snap-points-x') should be 100in repeat 42in. Was 100in repeat(42in).

pt along x axis with pixel repeat : 100pt repeat(42pt)
PASS cssRule.type is 1
PASS declaration.length is 1
FAIL declaration.getPropertyValue('-webkit-scroll-snap-points-x') should be 100pt repeat 42pt. Was 100pt repeat(42pt).

pixel/pixel destination : 10px 50px
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-destination') is '10px 50px'

pixel/percentage destination : 20px 40%
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-destination') is '20px 40%'

percentage/pixel destination : 0% 0px
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-destination') is '0% 0px'

percentage/percentage destination : 5% 100%
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-destination') is '5% 100%'

em/ex destination : 12em 16ex
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-destination') is '12em 16ex'

in/cm destination : 2in 5cm
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-destination') is '2in 5cm'

single pixel coordinate : 50px 100px
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-coordinate') is '50px 100px'

single percentage coordinate : 50% 100%
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-coordinate') is '50% 100%'

multiple pixel coordinates : 50px 100px 150px 100px 200px 100px
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-coordinate') is '50px 100px 150px 100px 200px 100px'

multiple percentage coordinates : 50% 100% 150% 100% 200% 100%
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-coordinate') is '50% 100% 150% 100% 200% 100%'

em/ex coordinate : 12em 16ex 16em 12ex
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-coordinate') is '12em 16ex 16em 12ex'

in/cm coordinate : 2in 5cm 5in 2cm
PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-scroll-snap-coordinate') is '2in 5cm 5in 2cm'
PASS successfullyParsed is true

TEST COMPLETE
